+++ CHANGES.rst
@@ -18,5 +18,8 @@ Features added
 Bugs fixed
 ----------
 
+* #12484: autodoc: Do not add an 'alias of ...' line using the :rst:dir:`autodata`
+  and :rst:dir:`autoattribute` directives for generic aliases having docstrings.
+
 Testing
 -------